[ChienBV] Refs #0110-01 - update code confirm email

+++ Web Gio Cha/Models/Define/GioHang.cs
@@ -7,9 +7,10 @@ namespace Web_Gio_Cha.Models.Define
 {
     public class GioHang
     {
-        public int TongSoLuong { get; set; }
+        public double TongSoLuong { get; set; }
         public decimal TongTien { get; set; }
 
         public IList<ItemGioHang> ItemGioHangList { get; set; }
+
     }
 }
